+++ src/topic/TopicMessageSubmitTransaction.js
@@ -140,27 +140,6 @@ export default class TopicMessageSubmitTransaction extends Transaction {
     }
 
     /**
-     * @param {TransactionId} transactionId
-     * @returns {this}
-     */
-    setTransactionId(transactionId) {
-        this._requireNotFrozen();
-
-        if (
-            transactionId.accountId == null ||
-            transactionId.validStart == null
-        ) {
-            throw new Error(
-                "`TopicMessageSubmitTransaction` does not support `TransactionId` built from `nonce`"
-            );
-        }
-
-        this._transactionIds.setList([transactionId]);
-
-        return this;
-    }
-
-    /**
      * @returns {?TopicId}
      */
     get topicId() {