+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * `TokenNftInfo.allowanceSpenderAccountId`
  * `ContractCreateFlow.sign[With]()`
 
+### Fixed
+
+ * `FileAppendTransaction` chunk size should default to 4096
+
 ### Deprecated
 
  * `AccountBalance.tokens` - Use the mirror node API https://docs.hedera.com/guides/docs/mirror-node-api/rest-api#api-v1-accounts instead