+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
  * `FileAppendTransaction` chunk size should default to 4096
+ * `ContractFunctionResult.gas` can bet `-1` when `null` in protobufs: #1208
 
 ### Fixed