+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
  * `FileAppendTransaction` chunk size should default to 4096
- * `ContractFunctionResult.gas` can bet `-1` when `null` in protobufs: #1208
+ * `ContractFunctionResult.gas` can be `-1` when `null` in protobufs: #1208
  * `FileAppendTransaction.setTransactionId()` not locking the transaction ID list
  * `TopicMessageSubmitTransaction.setTransactionId()` not locking the transaction ID list